+++ spec/controllers/mautic/connections_controller_spec.rb
@@ -27,7 +27,7 @@ module Mautic
 
     describe "GET #show" do
       it "returns a success response" do
-        get :show, params({ id: mautic_connection.to_param })
+        get :show, **params({ id: mautic_connection.to_param })
         expect(response).to be_successful
       end
     end
@@ -41,7 +41,7 @@ module Mautic
 
     describe "GET #edit" do
       it "returns a success response" do
-        get :edit, params({ id: mautic_connection.to_param })
+        get :edit, **params({ id: mautic_connection.to_param })
         expect(response).to be_successful
       end
     end
@@ -50,19 +50,19 @@ module Mautic
       context "with valid params" do
         it "creates a new MauticConnection" do
           expect {
-            post :create, params({ connection: FactoryBot.attributes_for(:mautic_connection) })
+            post :create, **params({ connection: FactoryBot.attributes_for(:mautic_connection) })
           }.to change(Connection, :count).by(1)
         end
 
         it "redirects to the created mautic_connection" do
-          post :create, params({ connection: FactoryBot.attributes_for(:mautic_connection) })
+          post :create, **params({ connection: FactoryBot.attributes_for(:mautic_connection) })
           expect(response).to be_successful
         end
       end
 
       context "with invalid params" do
         it "returns a success response (i.e. to display the 'new' template)" do
-          post :create, params({ connection: { url: "xxx" } })
+          post :create, **params({ connection: { url: "xxx" } })
           expect(response).to be_successful
         end
       end
@@ -72,20 +72,20 @@ module Mautic
       context "with valid params" do
 
         it "updates the requested mautic_connection" do
-          put :update, params({ id: mautic_connection.to_param, connection: { url: "https://newurl.com" } })
+          put :update, **params({ id: mautic_connection.to_param, connection: { url: "https://newurl.com" } })
           mautic_connection.reload
           expect(mautic_connection.url).to eq 'https://newurl.com'
         end
 
         it "redirects to the mautic_connection" do
-          put :update, params({ id: mautic_connection.to_param, connection: { url: "https://newurl.com" } })
+          put :update, **params({ id: mautic_connection.to_param, connection: { url: "https://newurl.com" } })
           expect(response).to redirect_to("/mautic/connections")
         end
       end
 
       context "with invalid params" do
         it "returns a success response (i.e. to display the 'edit' template)" do
-          put :update, params({ id: mautic_connection.to_param, connection: { url: "", secret: "" } })
+          put :update, **params({ id: mautic_connection.to_param, connection: { url: "", secret: "" } })
           expect(response).to be_successful
         end
       end
@@ -95,12 +95,12 @@ module Mautic
       it "destroys the requested mautic_connection" do
         mautic_connection # touch
         expect {
-          delete :destroy, params({ id: mautic_connection.to_param })
+          delete :destroy, **params({ id: mautic_connection.to_param })
         }.to change(Connection, :count).by(-1)
       end
 
       it "redirects to the mautic_connections list" do
-        delete :destroy, params({ id: mautic_connection.to_param })
+        delete :destroy, **params({ id: mautic_connection.to_param })
         expect(response).to have_http_status :redirect
       end
     end