+++ spec/controllers/mautic/webhook_controller_spec.rb
@@ -6,6 +6,7 @@ module Mautic
       webhook
     end
   end
+
   describe ApplicationController, type: :controller do
     controller FooController do
       def create
@@ -20,8 +21,8 @@ module Mautic
     it "receive request and find connection" do
 
       routes.draw { post "webhook/:mautic_connection_id" => "mautic/foo#create" }
-
-      post :create, params({ mautic_connection_id: oauth2 }.merge(JSON.load file_fixture("form_submit_webhook1.json")))
+      json = JSON.load file_fixture("form_submit_webhook1.json")
+      post :create, **params({ mautic_connection_id: oauth2 }.merge(json))
 
       expect(response).to be_successful
     end